requirements.txt
click==8.1.3
colorama==0.4.6
mypy-extensions==0.4.3
pathspec==0.10.1
platformdirs==2.5.2
PySide2==5.15.2.1
PyYAML==6.0
shiboken2==5.15.2.1
tomli==2.0.1